requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
# torch==1.12.1+cu113
# torchvision==0.13.1+cu113

# torch>=2.0.0
# torchvision>=0.15.0

# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu126

transformers==4.18.0
tensorboard==2.9.1
pytorch-lightning==1.9.1
timm==0.5.4
pycocotools==2.0.5
ninja==1.10.2
matplotlib==3.4.3
numpy==1.23.4
cython
wandb==0.18.3
scipy==1.13.1